Moved reduction from bin -> total_scattering.reduction
