+++ total_scattering/reduction/__init__.py
@@ -1,0 +1,1 @@
+from total_scattering.reduction.total_scattering_reduction import TotalScatteringReduction  # noqa
